Fixed control socket removal on reload.
The 'final' flag should be passed to desctructor in order to preserve
objects that must revive reload.

Some more complex logic is required to handle cases when socket name was
changed at reload - old one can be deleted, leaving as TODO.

+++ src/rie_control.c
@@ -218,7 +218,7 @@ rie_control_handle_socket_event(rie_control_t *ctl)
 
 
 void
-rie_control_delete(rie_control_t *ctl)
+rie_control_delete(rie_control_t *ctl, int final)
 {
     if (ctl == NULL) {
         return;
@@ -228,7 +228,9 @@ rie_control_delete(rie_control_t *ctl)
         (void) close(ctl->fd);
     }
 
-    (void) unlink(ctl->path);
+    if (final) {
+        (void) unlink(ctl->path);
+    }
 
     free(ctl);
 }
